geo-fetch-mcp/pyproject.toml
[project]
name = "geo-fetch-mcp"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "mcp>=1.0.0",
  "pydantic>=2.7.0",
  "orjson>=3.10.0",
  "python-dotenv>=1.0.1",
  "supabase>=2.6.0"
]


